further int test refactoring
