+++ src/it/resources/sample.conf
@@ -4,56 +4,9 @@ kinesis {
 
    application-name = "KinesisReliabilitySpec"
 
-   # The name of the this producer, we can have many producers per application.
-   # MUST contain the stream-name as a minimum. Any additional settings defined will override
-   # defaults in the kinesis.producer reference.conf for this producer only.
-   testProducer {
-      # The name of the producer stream
-      stream-name = "test-kinesis-reliability"
+   ### TODO To be removed...
 
-      # Can specify settings here as per default-producer, to override those defaults for this producer.
-
-      akka {
-         max-outstanding-requests = 50000
-      }
-
-      kpl {
-         Region = us-east-1
-      }
-   }
-
-   # The name of the consumer, we can have many consumers per application
-   testConsumer {
-      # The name of the consumer stream, MUST be specified per consumer
-      stream-name = "test-kinesis-reliability"
-
-      # Use localstack for integration test
-      kcl {
-         kinesisEndpoint = "https://localhost:4568"
-         dynamoDBEndpoint = "https://localhost:4569"
-
-         AWSCredentialsProvider = "com.weightwatchers.reactive.kinesis.common.TestCredentials|foo|bar"
-
-         # dramatically reduce default values.
-         # This will speed up the integration test by factor 20x or greater
-         maxRecords = 100
-         metricsLevel = NONE
-         failoverTimeMillis = 500
-         shardSyncIntervalMillis = 1000
-         idleTimeBetweenReadsInMillis = 100
-         parentShardPollIntervalMillis = 1000
-      }
-
-      worker {
-         batchTimeoutSeconds = 1
-         failedMessageRetries = 0
-         failureTolerancePercentage = 0
-         gracefulShutdownHook = false
-         shutdownTimeoutSeconds = 10
-      }
-   }
-
-   ## Test specific properties
+   ## (Simplexxx) Test specific properties
    test {
       # These settings basically say: Send 5,000,000 total messages, split into 2500 batches with 100 ms between batches
       # Equates to max throughput of 20,000/sec on producer
